Moved new functions to appropriate _lib locations
